For gemfiles/* use monthly interval
These shouldn't change as much (mostly with rails) but we aren't as interested in updating so frequently. Also issues with Dependabot downgrading dependencies that it shouldn't be so we want to review without auto merging for now. Just too hard to manage for ActiveAdmin due to the number of dependencies we declare.