Delete mergify.yml
Dependabot is downgrading dependencies it shouldn't be but its hard to manage with several gemfiles and all being auto merged. If it weren't for the wrong downgrading issue, I don't mind auto merging but its hard to manage so let's try it without for now. We'll leave it on for Arbre/Inherited Resources since they have far fewer dependencies compared to ActiveAdmin.